+++ tests/_plugin_tests/input_plugins/test_ascii_1d_profile_loader.py
@@ -28,7 +28,7 @@ import numpy as np
 import pytest
 
 from pydidas.contexts import ScanContext
-from pydidas.core import Dataset
+from pydidas.core import Dataset, UserConfigError
 from pydidas.core.utils import get_random_string
 from pydidas.data_io import export_data
 from pydidas.plugins import InputPlugin
@@ -77,6 +77,19 @@ def test_data():
 
 
 @pytest.fixture(scope="module")
+def test_data_multicol():
+    _x0_label = "; ".join([f"{i}: {get_random_string((6))}" for i in range(6)])
+    return Dataset(
+        np.random.random((145, 6)),
+        axis_labels=[_x_label, _x0_label],
+        axis_units=[_x_unit, ""],
+        axis_ranges=[_CUSTOM_XRANGE, np.arange(6)],
+        data_unit="some counts",
+        data_label="Test data",
+    )
+
+
+@pytest.fixture(scope="module")
 def stored_data(temp_path, test_data):
     for _ii in range(20):
         export_data(
@@ -95,10 +108,12 @@ def test__creation(plugin):
 
 
 def test_pre_execute(plugin):
+    plugin._ASCII1dProfileLoader__yslice = 1
+    plugin._ASCII1dProfileLoader__xscale_valid = True
     plugin.pre_execute()
-    assert plugin._standard_kwargs["forced_dimension"] == 1
     assert plugin._standard_kwargs["roi"] is None
     assert plugin._config["pre_executed"]
+    assert plugin._ASCII1dProfileLoader__yslice is None
     assert not plugin._ASCII1dProfileLoader__xscale_valid
 
 
@@ -126,6 +141,94 @@ def test_get_frame(setup_scan, stored_data, test_data, plugin, custom_xscale):
         assert np.all(_data.axis_ranges[0] == test_data.axis_ranges[0])
         assert _data.axis_labels[0] == test_data.axis_labels[0]
         assert _data.axis_units[0] == test_data.axis_units[0]
+
+
+@pytest.mark.parametrize("col_x", [None, 0, 2])
+@pytest.mark.parametrize("col_y", [None, 1, 4])
+@pytest.mark.parametrize("format", ["txt", "dat", "csv"])
+@pytest.mark.parametrize("header", [True, False])
+def test_get_frame__multicol__w_written_x_col(
+    setup_scan, test_data_multicol, temp_path, plugin, col_x, col_y, format, header
+):
+    export_data(
+        temp_path / f"test_multicol_00.{format}",
+        test_data_multicol,
+        overwrite=True,
+        write_header=header,
+    )
+    export_data(
+        temp_path / f"test_multicol_01.{format}",
+        test_data_multicol + 1,
+        overwrite=True,
+        write_header=header,
+    )
+    _stored_data = np.column_stack(
+        (test_data_multicol.axis_ranges[0], test_data_multicol)
+    )
+    SCAN.set_param_value("scan_name_pattern", f"test_multicol_##.{format}")
+    plugin.set_param_value("x_column", col_x)
+    plugin.set_param_value("y_column", col_y)
+    plugin.pre_execute()
+    if col_y is None:
+        with pytest.raises(UserConfigError):
+            _ = plugin.get_frame(0)
+        return
+    _ref_x = (
+        test_data_multicol.axis_ranges[0]
+        if (header and format in ["txt", "csv"]) or col_x == 0
+        else (
+            np.arange(test_data_multicol.shape[0])
+            if col_x is None
+            else test_data_multicol[:, col_x - 1]
+        )
+    )
+    _ref_y = _stored_data[:, col_y]
+    _data, _ = plugin.get_frame(0)
+    _data_1, _ = plugin.get_frame(1)
+    assert np.allclose(_data.axis_ranges[0], _ref_x)
+    assert np.allclose(_data, _ref_y)
+    assert np.allclose(_data_1, _ref_y + 1)
+
+
+@pytest.mark.parametrize("col_x", [None, 0, 2])
+@pytest.mark.parametrize("col_y", [1, 4])
+@pytest.mark.parametrize("format", ["txt", "dat", "csv"])
+def test_get_frame__multicol__no_written_x_col(
+    setup_scan, test_data_multicol, temp_path, plugin, col_x, col_y, format
+):
+    export_data(
+        temp_path / f"test_multicol_11.{format}",
+        test_data_multicol,
+        overwrite=True,
+        x_column=False,
+    )
+    SCAN.set_param_value("scan_name_pattern", f"test_multicol_11.{format}")
+    plugin.set_param_value("x_column", col_x)
+    plugin.set_param_value("y_column", col_y)
+    plugin.pre_execute()
+    _data, _ = plugin.get_frame(11)
+    if col_x is None:
+        assert np.allclose(_data.axis_ranges[0], np.arange(_data.size))
+    else:
+        assert np.allclose(_data.axis_ranges[0], test_data_multicol[:, col_x])
+    assert np.allclose(_data, test_data_multicol[:, col_y])
+
+
+def test_get_frame__w_custom_x_column(setup_scan, test_data, stored_data, plugin):
+    _x0, _x_delta = 42, 1.234
+    _x_label, _x_unit = get_random_string(6), get_random_string(3)
+    plugin.set_param_value("use_custom_xscale", True)
+    plugin.set_param_value("x0_offset", _x0)
+    plugin.set_param_value("x_delta", _x_delta)
+    plugin.set_param_value("x_label", _x_label)
+    plugin.set_param_value("x_unit", _x_unit)
+    plugin.pre_execute()
+    _data, _ = plugin.get_frame(11)
+    assert np.allclose(_data.axis_ranges[0], np.arange(_data.size) * _x_delta + _x0)
+    assert _data.axis_labels[0] == _x_label
+    assert _data.axis_units[0] == _x_unit
+    assert np.allclose(_data, test_data + 11)
+    assert plugin._ASCII1dProfileLoader__xscale_valid
 
 
 @pytest.mark.parametrize("ordinal", [0, 7])